fix: bug in youtube embeds
