+++ server/lib/notion/blocks/media/BlockEmbed.tsx
@@ -1,5 +1,7 @@
 import { GetBlockResponse } from "@notionhq/client/build/src/api-endpoints";
 import { renderToStaticMarkup } from "react-dom/server";
+import getYouTubeEmbedLink from "../../../parser/helpers/getYouTubeEmbedLink";
+import getYouTubeID from "../../../parser/helpers/getYouTubeID";
 import BlockHandler from "../../BlockHandler";
 
 export const BlockEmbed = (c: GetBlockResponse, handler: BlockHandler) => {
@@ -18,6 +20,11 @@ export const BlockEmbed = (c: GetBlockResponse, handler: BlockHandler) => {
           <a href={url}>{url}</a>
         </div>
       );
+    } 
+   
+    const yt = getYouTubeID(url);
+    if (yt) {
+      url = getYouTubeEmbedLink(yt);
     }
   }
   return renderToStaticMarkup(
