+++ server/lib/parser/helpers/getYouTubeEmbedLink.ts
@@ -1,0 +1,3 @@
+export default function getYouTubeEmbedLink(id: string): string {
+  return `https://www.youtube.com/embed/${id}?`.replace(/"/, '');
+}
